+++ java-fireworks.md
@@ -8,3 +8,7 @@
   name, description, price, loudnessRating;
 - On the firework show page, I want to see the font color changed based on the loudness rating of the firework.
   - Account for possible loudness ratings from 1 to 3.
+
+- On `/fireworks/new` there should be a form to add a new firework with all it's fields
+- After submitting the form, the new firework should be added to the list stored in session.
+- The user should be redirected back to `/fireworks`
